feat: implement data reading from external sources
IMPORTANT: not working as expected yet!

The application still does not check whether a transaction added from
raw CSV data is already applied.

Furthermore, the application does not correctly handle the stock splits
at this point. Splits are applied to ALL the previous transactions. They
should only be applied to the transactions active at that time.

Currently working on an improved way to handle this.